requirements.txt
# Core Django Dependencies
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# Authentication & Security
PyJWT==2.8.0
cryptography==41.0.7
django-oauth-toolkit==1.7.1
djangorestframework-simplejwt==5.3.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
django-allauth==0.57.0
django-ratelimit==4.1.0
django-phonenumber-field==7.2.0
phonenumbers==8.13.25

# Database
psycopg2-binary==2.9.9
pymongo==4.6.0
redis==5.0.1

# Google Cloud
google-cloud-secret-manager==2.16.4
google-cloud-storage==2.10.0
google-cloud-run==0.10.5

# AI/ML Governance
openai==1.3.5
anthropic==0.7.7
langchain==0.0.340
tiktoken==0.5.1

# Monitoring & Observability
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk==1.38.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
responses==0.24.1

# Code Quality
black==23.11.0
flake8==6.1.0
isort==5.12.0
bandit==1.7.5
safety==3.0.1

# API Documentation
drf-spectacular==0.26.5
